Crate ibc_client_tendermint_types

ICS-07: Tendermint Client implements a client verification algorithm for blockchains which use the Tendermint consensus algorithm.

Modules

error

Defines the Tendermint light client’s error type

proto

Re-exports ICS-07 Tendermint light client from ibc-proto crate.

Structs

AllowUpdate

ClientState

Defines data structure for Tendermint client state.

ConsensusState

Defines the Tendermint light client’s consensus state

Header

Tendermint consensus header

Misbehaviour

Tendermint light client’s misbehaviour type

TrustThreshold

TrustThreshold defines the level of trust that a client has towards a set of validators of a chain.

Constants

TENDERMINT_CLIENT_STATE_TYPE_URL

TENDERMINT_CLIENT_TYPE

TENDERMINT_CONSENSUS_STATE_TYPE_URL

TENDERMINT_HEADER_TYPE_URL

TENDERMINT_MISBEHAVIOUR_TYPE_URL

Functions

client_type

Returns the tendermint ClientType